add notebook plotting period values

+++ szegedy.py
@@ -16,26 +16,26 @@ class SzegedyRandomWalk:
     def __construct_base(self):
         base = []
         for i in range(self.n):
-            base_i = np.zeros(self.n)
+            base_i = np.zeros(self.n, dtype=np.float128)
             base_i[i] = 1
             base.append(base_i)
-        return np.asarray(base)
+        return np.asarray(base, dtype=np.float128)
 
     def _calculate_alphas(self):
         alphas = []
         for i in range(self.n):
-            alpha_i = np.zeros(self.n * self.n)
+            alpha_i = np.zeros(self.n * self.n, dtype=np.float128)
             for j in range(self.n):
-                alpha_i += np.sqrt(self.markov_chain[i, j]) * np.kron(self.base[i], self.base[j])
+                alpha_i += np.sqrt(self.markov_chain[i, j], dtype=np.float128) * np.kron(self.base[i], self.base[j])
             alphas.append(alpha_i)
         return alphas
 
     def _calculate_betas(self):
         betas = []
         for i in range(self.n):
-            beta_i = np.zeros(self.n * self.n)
+            beta_i = np.zeros(self.n * self.n, dtype=np.float128)
             for j in range(self.n):
-                beta_i += np.sqrt(self.markov_chain[i, j]) * np.kron(self.base[j], self.base[i])
+                beta_i += np.sqrt(self.markov_chain[i, j], dtype=np.float128) * np.kron(self.base[j], self.base[i])
             betas.append(beta_i)
         return betas
 
@@ -43,8 +43,8 @@ class SzegedyRandomWalk:
         return np.column_stack(tuple(self.alphas)), np.column_stack(tuple(self.betas))
 
     def _construct_refs(self):
-        ref1 = 2 * self.A @ np.conjugate(self.A).T - np.identity(self.n * self.n)
-        ref2 = 2 * self.B @ np.conjugate(self.B).T - np.identity(self.n * self.n)
+        ref1 = 2 * self.A @ np.conjugate(self.A, dtype=np.float128).T - np.identity(self.n * self.n, dtype=np.float128)
+        ref2 = 2 * self.B @ np.conjugate(self.B, dtype=np.float128).T - np.identity(self.n * self.n, dtype=np.float128)
         return ref1, ref2
 
     def _construct_evolution_operator(self):
